refactor(bff): streamline user service by using session data
This commit refactors the user module to improve separation of concerns and reduce redundant code.

The UserService no longer fetches session information from the cache directly.
Instead, it receives the complete sessionData object from the UserController, which in turn gets
it from the validateSession-middleware.

This change centralizes session handling within the middleware, making the UserService cleaner and
more focused on its primary responsibility of communicating with the backend user API.