refactor(bff): simplify note service by using session data
This commit refactors the note module to improve separation of concerns and reduce redundant code.

The NoteService no longer fetches session information from the cache directly.
Instead, it receives the complete sessionData object from the NoteController, which in turn gets
it from the validateSession-middleware.

This change centralizes session handling within the middleware, making the NoteService cleaner
and more focused on its primary responsibility of communicating with the backend note API.